pyproject.toml
[tool.poetry]
name = "pyhelpers"
version = "0.1.0"
description = ""
authors = ["RodionfromHSE"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"
requests = "^2.31.0"
notebook = "^7.1.1"
pandas = "^2.2.1"
tqdm = "^4.66.2"
click = "^8.1.7"
openai = "^1.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
